+++ thread_rover.c
@@ -72,7 +72,7 @@ static int threadRoverSelectLoop(Runtime* _runtime, RoverOutput* _rover, DriverO
 void* threadRover(void* _arg)
 {
   int res = 0;
-  int exit_code = 0;
+  intptr_t exit_code = 0;
   Runtime* runtime = (Runtime*)_arg;
   RoverOutput* rover;
   DriverOutput* driver;
